# [Coro] Amortize debug info processing cost in CoroSplit