More details about this stack are [in this topic on discourse](https://discourse.llvm.org/t/rfc-amortizing-debug-info-processing-cost-in-corosplit/81275). 

TL;DR: In large modules with a bunch of coroutines the cost of debug info metadata processing can get very high (e.g. adding over a minute to the compile time). This stack does some refactoring around function cloning to significantly speed up the pass.

This makes CoroSplit pass almost as fast with -g2 as it is with -g1 on
the sample cpp file used with other parts of this stack:

|                 | Baseline | IdentityMD set | Prebuilt CommonDI | MetadataPred (fin) |
|-----------------|----------|----------------|-------------------|--------------------|
| CoroSplitPass   | 306ms    | 221ms          | 68ms              | 3.8ms              |
| CoroCloner      | 101ms    | 72ms           | 0.5ms             | 0.5ms              |
| CollectCommonDI | -        | -              | 63ms              | -                  |
| Speed up        | 1x       | 1.4x           | 4.5x              | 80x                |

Each commit is individually buildable and reviewable. I can extract them into separate PRs if the overall direction makes sense. 

-----

The final version has been merged in a series of PRs (❗️marks the salient ones):
1. #112948
2. #112976
3. #114537
4. #118620
5. #118621
6. #118622
7. #118623 
8. #118624
9. #118625
10. ❗️ #118627
11. #118628
12. #129143
13. #129144
14. #129145
15. #129146
16. ❗️ #129147
17. ❗️ #129148
18. #129149
19. #129150
20. #129151
21. #129152
22. #129153
23. #129154