@artempyanykh 2024-10-01
Thanks @felipepiovezan! There's some commentary around the commits in [this discourse topic](https://discourse.llvm.org/t/rfc-amortizing-debug-info-processing-cost-in-corosplit/81275), in case it's useful.