@felipepiovezan 2024-10-17
Hi @artempyanykh, while I follow Adrian's suggestions, you could extract the first commit into a separate PR and we'll start reviwing/merging the NFC stuff one by one