@artempyanykh 2024-10-18
@felipepiovezan I extracted the first 2 commits into #112948 and #112976.