@adrian-prantl 2024-11-11
> With https://github.com/llvm/llvm-project/pull/114537 out, there’s quite a bit of sequencing in the remaining 12 commits. I can combine some of the remaining commits into larger PRs if it makes the review logistics a bit easier.

On the contrary, 12 PRs with a trivial NFC refactor will be easier to review than 1 large one.