@artempyanykh 2024-11-20
Rebased on trunk