@artempyanykh 2024-11-20
> On the contrary, 12 PRs with a trivial NFC refactor will be easier to review than 1 large one.

I agree. Just being mindful of sequencing and iteration time. After the review of #114537 is complete, I'll try using [spr](https://github.com/spacedentist/spr) or [stack-pr](https://github.com/modularml/stack-pr) to parallelize the rest of the stack a bit.