@artempyanykh 2024-12-04
Thanks for checking @felipepiovezan! I sent the rest of the stack as PRs via [stack-pr](https://github.com/modularml/stack-pr): the root is at #118620. They still need to land in order, but could be reviewed separately/concurrently. Bottom half of the stack are refactorings. More interesting changes start at #118627.