@artempyanykh 2025-04-21
This has now been fully merged. I updated the description with references to relevant PRs. The final implementation turned out faster than the initial one: in terms of the CoroSplit pass runtime on a sample cpp file with `-g2`: 306ms -> 3.8ms @ 80x now vs 306ms -> 17ms @ 18x which is almost as fast as `-g1`.

Thanks @felipepiovezan, @TylerNowicki, and @adrian-prantl for reviews!